Alphabetized, Fixed Errors, Separated Terrains
I put this in alpha order to help us keep track of what we add. Try to keep it in alpha order please. xD (I couldn't be bothered to alphabetize the concealed power types though.)
I can foresee problems with the terrain moves so I separated them for easier access.